# backend/docs/swagger.yaml
basePath: /
definitions:
  ArticleListResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/ArticleResponse'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  ArticleResponse:
    properties:
      article_type:
        enum:
        - markdown
        - external
        example: markdown
        type: string
      author:
        $ref: '#/definitions/AuthorResponse'
      content:
        example: 記事の本文です...
        type: string
      created_at:
        example: "2026-01-06T12:00:00Z"
        type: string
      department:
        enum:
        - Dev
        - MKT
        - Ops
        example: Dev
        type: string
      description:
        example: 記事の要約や説明文です
        type: string
      external_url:
        example: https://example.com/article
        type: string
      id:
        example: 1
        type: integer
      slug:
        example: go-api-development
        type: string
      status:
        enum:
        - draft
        - internal
        - public
        example: public
        type: string
      tags:
        example:
        - Go
        - Backend
        - Echo
        items:
          type: string
        type: array
      thumbnail_url:
        example: https://example.com/thumbnail.jpg
        type: string
      title:
        example: Go言語でのAPI開発入門
        type: string
      updated_at:
        example: "2026-01-06T12:00:00Z"
        type: string
    type: object
  AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/UserResponse'
    type: object
  AuthenticateRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  AuthorResponse:
    properties:
      affiliation:
        example: 開発部
        type: string
      icon_url:
        example: https://example.com/icon.jpg
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 山田太郎
        type: string
    type: object
  CreateExternalArticleRequest:
    properties:
      department:
        enum:
        - 開発
        - マーケティング
        - 組織管理
        example: 開発
        type: string
      status:
        enum:
        - draft
        - internal
        - public
        example: public
        type: string
      url:
        example: https://zenn.dev/example/articles/sample
        type: string
    required:
    - department
    - status
    - url
    type: object
  ErrorResponse:
    properties:
      error:
        example: エラーメッセージ
        type: string
      message:
        example: 詳細なエラー情報
        type: string
    type: object
  SignUpRequest:
    properties:
      affiliation:
        enum:
        - 開発
        - マーケティング
        - 組織管理
        example: 開発
        type: string
      email:
        example: user@example.com
        type: string
      name:
        example: 山田太郎
        type: string
      password:
        example: password123
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  UserDetailResponse:
    properties:
      affiliation:
        example: Dev
        type: string
      articles:
        items:
          $ref: '#/definitions/ArticleResponse'
        type: array
      created_at:
        example: "2026-01-01T12:00:00Z"
        type: string
      icon_url:
        example: https://example.com/icon.jpg
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 山田太郎
        type: string
      portfolio_key:
        example: abc123xyz
        type: string
    type: object
  UserResponse:
    properties:
      affiliation:
        example: 開発部
        type: string
      email:
        example: user@example.com
        type: string
      icon_url:
        example: https://example.com/icon.jpg
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 山田太郎
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    チーム1のテックブログシステムのためのAPI仕様書です。
    記事の取得、閲覧などの機能を提供します。
  title: Team1 Blog API
  version: "1.0"
paths:
  /api/articles:
    get:
      consumes:
      - application/json
      description: 公開されているブログ記事の一覧を取得します。ログイン済みの場合は内部公開記事も含まれます。ページネーション、部署フィルタ、ステータスフィルタをサポートしています。
      parameters:
      - default: 1
        description: 'ページ番号 (デフォルト: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: '1ページあたりの件数 (デフォルト: 10, 最大: 100)'
        in: query
        name: limit
        type: integer
      - description: 部署でフィルタ (Dev, MKT, Ops)
        enum:
        - Dev
        - MKT
        - Ops
        in: query
        name: department
        type: string
      - description: ステータスでフィルタ (internal, public, all)。ゲストの場合は常にpublicのみ
        enum:
        - internal
        - public
        - all
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 記事一覧
          schema:
            $ref: '#/definitions/ArticleListResponse'
        "400":
          description: リクエストパラメータが不正です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: サーバー内部でエラーが発生しました
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: 記事一覧を取得
      tags:
      - 記事 (Articles)
  /api/articles/{slug}:
    get:
      consumes:
      - application/json
      description: 指定されたslugのブログ記事の詳細を取得します。内部公開記事の場合はログインが必要です。
      parameters:
      - description: 記事のスラグ
        example: '"go-api-development"'
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 記事詳細
          schema:
            $ref: '#/definitions/ArticleResponse'
        "403":
          description: 内部公開記事にアクセスするにはログインが必要です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: 記事が見つかりません
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: サーバー内部でエラーが発生しました
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: 記事詳細を取得
      tags:
      - 記事 (Articles)
  /api/articles/external:
    post:
      consumes:
      - application/json
      description: 外部サイト（Zenn、Noteなど）の記事URLを指定して、OGP情報を取得し記事として登録します。
      parameters:
      - description: 外部記事作成リクエスト
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateExternalArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 作成された記事
          schema:
            $ref: '#/definitions/ArticleResponse'
        "400":
          description: リクエストが不正です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: 認証が必要です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: サーバー内部でエラーが発生しました
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: 外部記事を作成
      tags:
      - 記事 (Articles)
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 既存のユーザーを認証し、新しい認証トークンを発行します。
      parameters:
      - description: ユーザー情報 (メールアドレスとパスワード)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AuthenticateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 認証成功。新しい認証トークンを返します。
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: リクエストボディが不正です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: 認証に失敗しました (メールアドレスまたはパスワードが正しくありません)
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: サーバー内部でエラーが発生しました
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: ログイン (Log In)
      tags:
      - 認証 (Auth)
  /api/auth/logout:
    post:
      description: クッキーを削除してログアウトします。
      produces:
      - application/json
      responses:
        "200":
          description: ログアウト成功
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ログアウト (Logout)
      tags:
      - 認証 (Auth)
  /api/auth/me:
    get:
      description: Cookieからトークンを読み取り、現在ログイン中のユーザー情報を返します。
      produces:
      - application/json
      responses:
        "200":
          description: ユーザー情報
          schema:
            $ref: '#/definitions/UserResponse'
        "401":
          description: 認証されていません
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: 現在のユーザー情報を取得
      tags:
      - 認証 (Auth)
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: 新しいユーザーアカウントを作成し、認証トークンとユーザー情報を返します。
      parameters:
      - description: ユーザー情報 (メールアドレス、パスワード、名前)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 登録成功。ユーザー情報と認証トークンを返します。
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: リクエストボディが不正です
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: 指定されたメールアドレスは既に使用されています
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: サーバー内部でエラーが発生しました
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: 新規ユーザー登録 (Sign Up)
      tags:
      - 認証 (Auth)
  /users/{id}:
    get:
      consumes:
      - application/json
      description: 指定されたIDのユーザー公開プロフィール（名前、所属、アイコン、記事一覧）を取得します。未認証でもアクセス可能ですが、外部公開記事のみ表示されます。
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: ポートフォリオキー（内部公開記事も表示する場合に必要）
        in: query
        name: portfolio_key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: ユーザー詳細取得
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: '認証トークンを''Bearer ''に続けて入力してください。 (例: Bearer {JWTトークン})'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"